backend/requirements.txt
fastapi==0.110.1
uvicorn==0.25.0
motor==3.3.1
pymongo==4.5.0
pydantic>=2.6.4
email-validator>=2.2.0
python-jose>=3.3.0
requests>=2.31.0
pytest>=8.0.0
